inventory-service-python/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
ariadne==0.20.0
mysql-connector-python==8.2.0
python-dotenv==1.0.0
httpx==0.25.2
pydantic==2.5.2



